backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "openai-whisper>=20231117",
    "lmstudio>=1.5.0",
    "numpy>=1.26.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
]
